fix RemoveFromOrgModal
